Added belt clamp corner and a simple benchdog.
Belt clamp corner is actually a half of the corner, intended for gluing two
such halves together. It just prints better this way (without any supports).

Benchdog needs a square stick insert to work. A simple 3D-printed dowel with
a cut-off didn't work well, because of the cutting forces working in the same
plane as material layers.